noir_artifact_cli

Module execution

Structs

ExecutionResults

Results of a circuit execution.

ReturnValues

The decoded return witnesses.

Functions

append_input_value_to_string 🔒

check_witness

Compare return values to expectations, returning errors if something unexpected was returned.

execute

Execute a circuit and return the output witnesses.

input_value_to_string

save_and_check_witness

Print some information and save the witness if an output directory is specified, then checks if the expected return values were the ones we expected.

save_witness

Save the witness stack to a file.

show_diagnostic

Print an error stack trace, if possible.