noir_artifact_cli::fs
Module artifact

Functions

read_bytecode_from_file

Returns the circuit’s bytecode read from the file at the given location

read_program_from_file

Read a ProgramArtifact. Returns error if it turns out to be a ContractArtifact.

save_build_artifact_to_file 🔒

save_contract_to_file

save_program_to_file

write_to_file

Create the parent directory if needed and write the bytes to a file.